requirements.txt
flask
flask-cors
groq
pymongo[srv]
python-dotenv
dnspython
